Move NodeRelation classes to their own files to prevent circular imports
